Move axes classes to separate files and *rename*
